started on makefiles and rewrote winnowing
winnowing is now entirely functional, with a queue module that I defined (I will add specs for that module later).

also added the scripts/files we need for the make commands, although they do not work yet because things are not finalized. I recommend making a main.ml file and putting the interface there instead